#6 [chore] 프로젝트명 변경
