--- src/main/java/org/sopt/cerdeuk_server/CerdeukServerApplication.java
@@ -1,13 +1,0 @@
-package org.sopt.cerdeuk_server;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class CerdeukServerApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(CerdeukServerApplication.class, args);
-	}
-
-}
