--- src/main/java/org/sopt/cerdeuk_server/global/error/exception/InvalidValueException.java
@@ -1,9 +1,0 @@
-package org.sopt.cerdeuk_server.global.error.exception;
-
-import org.sopt.cerdeuk_server.global.error.code.ErrorCode;
-
-public class InvalidValueException extends BusinessException {
-	public InvalidValueException(ErrorCode errorCode) {
-		super(errorCode);
-	}
-}
