--- src/main/java/org/sopt/cerdeuk_server/global/error/exception/NotFoundException.java
@@ -1,9 +1,0 @@
-package org.sopt.cerdeuk_server.global.error.exception;
-
-import org.sopt.cerdeuk_server.global.error.code.ErrorCode;
-
-public class NotFoundException extends BusinessException {
-	public NotFoundException(final ErrorCode errorCode) {
-		super(errorCode);
-	}
-}
