+++ src/main/java/org/sopt/certi_server/CertiServerApplication.java
@@ -1,0 +1,13 @@
+package org.sopt.certi_server;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class CertiServerApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(CertiServerApplication.class, args);
+	}
+
+}
