+++ src/main/java/org/sopt/certi_server/global/error/exception/BusinessException.java
@@ -1,0 +1,16 @@
+package org.sopt.certi_server.global.error.exception;
+
+
+import org.sopt.certi_server.global.error.code.ErrorCode;
+
+import lombok.Getter;
+
+@Getter
+public class BusinessException extends RuntimeException {
+	private final ErrorCode errorCode;
+
+	public BusinessException(ErrorCode errorCode) {
+		super(errorCode.getMessage());
+		this.errorCode = errorCode;
+	}
+}
