+++ src/main/java/org/sopt/certi_server/global/error/exception/NotFoundException.java
@@ -1,0 +1,9 @@
+package org.sopt.certi_server.global.error.exception;
+
+import org.sopt.certi_server.global.error.code.ErrorCode;
+
+public class NotFoundException extends BusinessException {
+	public NotFoundException(final ErrorCode errorCode) {
+		super(errorCode);
+	}
+}
