+++ src/main/java/org/sopt/certi_server/global/error/exception/UnauthorizedException.java
@@ -1,0 +1,14 @@
+package org.sopt.certi_server.global.error.exception;
+
+import org.sopt.certi_server.global.error.code.ErrorCode;
+
+public class UnauthorizedException extends BusinessException {
+	public UnauthorizedException() {
+		super(ErrorCode.UNAUTHORIZED);
+	}
+
+	public UnauthorizedException(ErrorCode errorCode) {
+		super(errorCode);
+	}
+
+}
