+++ src/main/java/org/sopt/certi_server/global/jwt/repository/TokenRepository.java
@@ -1,0 +1,12 @@
+package org.sopt.certi_server.global.jwt.repository;
+
+import java.util.Optional;
+
+import org.sopt.certi_server.global.jwt.entity.Token;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface TokenRepository extends CrudRepository<Token, Long> {
+	Optional<Token> findByRefreshToken(String refreshToken);
+}
