--- src/test/java/org/sopt/cerdeuk_server/CerdeukServerApplicationTests.java
@@ -1,13 +1,0 @@
-package org.sopt.cerdeuk_server;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class CerdeukServerApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
