+++ src/test/java/org/sopt/certi_server/CertiServerApplicationTests.java
@@ -1,0 +1,13 @@
+package org.sopt.certi_server;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CertiServerApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
